labs/mem_protect: Move stack back to local core
Loads and stores are routed on separate NoCs. This means that a store
can arrive after a subsequent load. So if the stack is placed anywhere
but on the local core there will exist a race condition that must be
taken into account, which is a complex task.
Instead move the stack back to local core SRAM.

* labs/mem_protect/src/mpf_isr.s: Remove.
* labs/mem_protect/src/emain.c (void mpf_isr): Replace above
assembler page fault handler with C version.
(int main): Simplify. Do not test stack page.
* labs/mem_protect/src/main.c: Simplify.
* labs/mem_protect/build.sh: Don't link with removed assembly
mpf isr. Use linker script that will place stack in local core
SRAM.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>