labs/hardware_barrier: Fix race condition
The hardware barrier example would fail occasionally because of a race
condition. This commit adds a synchronization flag between the first
core in the group and the host. When the host sets the flag core0 will
busy-wait until the flag is cleared. This means that the other cores
will be idling waiting for the WAND interrupt. But that won't happen
until all cores in the group issue the WAND instruction. This creates a
window for the host to read all cores' iteration counters.

This commit also removes old unused ISR assembly code that predates
Epiphany GCC's interrupt attribute.

* labs/hardware_barrier/src/emain.c (int main): Add
synchronization point between first E core and host so that
the host can read out the core iteration counters in a
race-free manner.
* labs/hardware_barrier/src/main.c: Likewise.
* labs/hardware_barrier/src/hwb_isr.s: Remove.
* labs/hardware_barrier/build.sh: Remove all references to
hwb_isr.s.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>