tests/test-linker-script-symbols: Add linker script symbols regression test
In ESDK 2016.3 Epiphany GCC USER_LABEL_PREFIX changed from "_" to "",
(Epiphany GCC 5+), which means we should adjust the linker scripts to provide
the same names from C for symbols that are part of the ABI(?). Some
people use it and there's no reason to break code compatibility.

The test checks that these symbols are present:
_PROG_SIZE_FOR_CORE_
_HEAP_SIZE_FOR_CORE_
_HALF_BANK_SIZE_
_NUM_ROWS_IN_CHIP_
_NUM_COLS_IN_CHIP_
_FIRST_CORE_ROW_
_FIRST_CORE_COL_
_CORE_ROW_
_CORE_COL_
_CORE_NUM_
_SHARED_DRAM_

Signed-off-by: Ola Jeppsson <ola@adapteva.com>